# front/src/app/api/models/RegisterDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
 
export type RegisterDto = {
    firstName: string;
    lastName: string;
    email: string;
    password: string;
};