# front/src/app/api/models/UpdatePersonDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
 
export type UpdatePersonDto = {
    firstName: string;
    lastName: string;
    email: string;
    updatedAt: string;
};